+++ labelbox/data/annotation_types/classification/classification.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List, Union
+from typing import Any, Dict, List, Union, Optional
 
 try:
     from typing import Literal
@@ -24,32 +24,47 @@ class ClassificationAnswer(FeatureSchema):
     - Represents a classification option.
     - Because it inherits from FeatureSchema
         the option can be represented with either the name or feature_schema_id
+
+    - The key frame arg only applies to video classifications.
+      Each answer can have a key frame indepdent of the others.
+        So unlike object annotations, classification annotations
+          track key frames at a classification answer level.
     """
     extra: Dict[str, Any] = {}
+    keyframe: Optional[bool] = None
 
-
-class VideoClassificationAnswer(ClassificationAnswer):
-    """
-    Each answer can have a key frame indepdent of the others.
-    So unlike object annotations, classification annotations
-      track key frames at a classification answer level.
-    """
-    keyframe: bool
+    def dict(self, *args, **kwargs):
+        res = super().dict(*args, **kwargs)
+        if res['keyframe'] is None:
+            res.pop('keyframe')
+        return res
 
 
 class Radio(BaseModel):
-    """ A classification with only one selected option allowed """
-    answer: Union[VideoClassificationAnswer, ClassificationAnswer]
+    """ A classification with only one selected option allowed
+
+    >>> Radio(answer = ClassificationAnswer(name = "dog"))
+
+    """
+    answer: ClassificationAnswer
 
 
 class Checklist(_TempName):
-    """ A classification with many selected options allowed """
+    """ A classification with many selected options allowed
+
+    >>> Checklist(answer = [ClassificationAnswer(name = "cloudy")])
+
+    """
     name: Literal["checklist"] = "checklist"
-    answer: Union[VideoClassificationAnswer, ClassificationAnswer]
+    answer: ClassificationAnswer
 
 
 class Text(BaseModel):
-    """ Free form text """
+    """ Free form text
+
+    >>> Text(answer = "some text answer")
+
+    """
     answer: str
 
 
@@ -59,4 +74,4 @@ class Dropdown(_TempName):
     - This is not currently compatible with MAL.
     """
     name: Literal["dropdown"] = "dropdown"
-    answer: List[Union[VideoClassificationAnswer, ClassificationAnswer]]
+    answer: List[ClassificationAnswer]