+++ labelbox/data/serialization/labelbox_v1/classification.py
@@ -1,4 +1,3 @@
-from labelbox.data.annotation_types.classification.classification import VideoClassificationAnswer
 from typing import List, Union
 
 from pydantic.main import BaseModel
@@ -10,33 +9,23 @@ from ...annotation_types.types import Cuid
 
 
 class LBV1ClassificationAnswer(LBV1Feature):
-    ...
-
-
-def get_classification_answer(
-    answer: LBV1ClassificationAnswer
-) -> Union[ClassificationAnswer, VideoClassificationAnswer]:
-    kwargs = dict(feature_schema_id=answer.schema_id,
+    def to_common(self, answer: "LBV1ClassificationAnswer") -> ClassificationAnswer:
+        return ClassificationAnswer(feature_schema_id=answer.schema_id,
                   name=answer.title,
+                  keyframe = answer.keyframe,
                   extra={
                       'feature_id': answer.feature_id,
                       'value': answer.value
                   })
-    if answer.keyframe is not None:
-        return VideoClassificationAnswer(keyframe=answer.keyframe, **kwargs)
-    else:
-        return ClassificationAnswer(**kwargs)
 
-
-def get_lbv1_classification_answer(answer: Union[ClassificationAnswer,
-                                                 VideoClassificationAnswer]):
-    return LBV1ClassificationAnswer(
+    @classmethod
+    def from_common(cls, answer: ClassificationAnnotation) -> "LBV1ClassificationAnswer":
+        return cls(
         schema_id=answer.feature_schema_id,
         title=answer.name,
         value=answer.extra.get('value'),
         feature_id=answer.extra.get('feature_id'),
-        keyframe=getattr(answer, 'keyframe',
-                         None)  # Only applies to video clasifications
+        keyframe=answer.keyframe
     )
 
 
@@ -44,13 +33,13 @@ class LBV1Radio(LBV1Feature):
     answer: LBV1ClassificationAnswer
 
     def to_common(self) -> Radio:
-        return Radio(answer=get_classification_answer(self.answer))
+        return Radio(answer=self.answer.to_common())
 
     @classmethod
     def from_common(cls, radio: Radio, feature_schema_id: Cuid,
                     **extra) -> "LBV1Radio":
         return cls(schema_id=feature_schema_id,
-                   answer=get_lbv1_classification_answer(radio.answer),
+                   answer=LBV1ClassificationAnswer.from_common(radio.answer),
                    **extra)
 
 
@@ -59,7 +48,7 @@ class LBV1Checklist(LBV1Feature):
 
     def to_common(self) -> Checklist:
         return Checklist(answer=[
-            get_classification_answer(answer) for answer in self.answers
+            answer.to_common() for answer in self.answers
         ])
 
     @classmethod
@@ -67,7 +56,7 @@ class LBV1Checklist(LBV1Feature):
                     **extra) -> "LBV1Checklist":
         return cls(schema_id=feature_schema_id,
                    answers=[
-                       get_lbv1_classification_answer(answer)
+                       LBV1ClassificationAnswer.from_common(answer)
                        for answer in checklist.answer
                    ],
                    **extra)
@@ -78,7 +67,7 @@ class LBV1Dropdown(LBV1Feature):
 
     def to_common(self) -> Dropdown:
         return Dropdown(answer=[
-            get_classification_answer(answer) for answer in self.answer
+            answer.to_common() for answer in self.answer
         ])
 
     @classmethod
@@ -86,7 +75,7 @@ class LBV1Dropdown(LBV1Feature):
                     **extra) -> "LBV1Dropdown":
         return cls(schema_id=feature_schema_id,
                    answer=[
-                       get_lbv1_classification_answer(answer)
+                       LBV1ClassificationAnswer.from_common(answer)
                        for answer in dropdown.answer
                    ],
                    **extra)